Move event related code into its own module
It was weird having event related stuff inside `util`, which is
normally used for one-off functions.
